C1/pi4 audio/requirements.txt
astroid==3.3.5
autopep8==2.3.1
cffi==1.17.1
dill==0.3.9
isort==5.13.2
mccabe==0.7.0
numpy==2.1.3
paho-mqtt==2.1.0
platformdirs==4.3.6
PyAudio==0.2.14
pycodestyle==2.12.1
pycparser==2.22
pydub==0.25.1
pylint==3.3.1
scipy==1.14.1
simpleaudio==1.0.4
sounddevice==0.5.1
soundfile==0.12.1
tomlkit==0.13.2
